factored out check clause sp
