+++ wiggle_query_language/clauses/parsing_helpers/parse_statement_checks.py
@@ -1,6 +1,5 @@
 from re import split
 
-
 from exceptions.wql.parsing import (
     ClauseSyntaxError,
     IllegalCharacterError,
@@ -8,10 +7,10 @@ from exceptions.wql.parsing import (
     ParamSyntaxError,
     RelationshipNameSyntaxError,
 )
+from models.wql.enums.clauses import Clause
 from wiggle_query_language.clauses.regexes.make.make_patterns import (
     MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX,
 )
-
 from wiggle_query_language.clauses.regexes.patterns.properties import (
     CHECK_PROPERTIES_SYNTAX_REGEX,
     PROPERTIES_LIST_VALUE_REGEX,
@@ -20,6 +19,31 @@ from wiggle_query_language.clauses.regexes.patterns.relationships import (
     RELATIONSHIP_DIR_CHECK_REGEX,
 )
 from wiggle_query_language.clauses.regexes.patterns.utilities import ILLEGAL_CHARS_REGEX
+
+
+def check_clause_spelling(query_string: str, clause: Clause) -> bool:
+    """
+    Checks the MAKE for a spelling mistake.
+    :param query_string: The extracted MAKE statements.
+    :param clause: The extracted MAKE statements.
+    :return: True or raises and exception.
+    """
+
+    match clause:
+        case Clause.MAKE:
+            clause_syntax_regex = MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX
+        case Clause.FIND:
+            clause_syntax_regex = None
+        case _:
+            raise Exception()
+
+    if matches := clause_syntax_regex.findall(query_string):
+        for match in matches:
+            raise ClauseSyntaxError(
+                message=f"SyntaxError: {match} was not recognised did you mean {clause.value}?"
+            )
+
+    return True
 
 
 def check_property_syntax(params_string: str) -> bool:
@@ -68,22 +92,6 @@ def check_node_rel_properties(stmt_matches: list[str]) -> True:
                         raise ParamSyntaxError(
                             message=f"SyntaxError: {params_lists} missing a comma?"
                         )
-
-    return True
-
-
-def check_make_clause_spelling(query_string: str) -> bool:
-    """
-    Checks the MAKE for a spelling mistake.
-    :param query_string: The extracted MAKE statements.
-    :return: True or raises and exception.
-    """
-
-    if matches := MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX.findall(query_string):
-        for match in matches:
-            raise ClauseSyntaxError(
-                message=f"SyntaxError: {match} was not recognised did you mean MAKE?"
-            )
 
     return True
